- Change Company Name
- Change company name from DeadFish Studio to Îakaré Software'oka
