+++ .github/workflows/build.yml
@@ -1,4 +1,4 @@
-name: DeadFishStudio.CoreLibraryCI
+name: CoreLibraryCI
 on:
   push:
     branches: [ main ]
@@ -14,7 +14,7 @@ jobs:
     - name: Setup .NET
       uses: actions/setup-dotnet@v1
       with:
-        dotnet-version: 6.0.x
+        dotnet-version: 5.0.x
     - name: Restaurando pacotes
       run: dotnet restore
     - name: Construindo projetos
