+++ CoreLibrary.UnitTest/EErroEnum.cs
@@ -1,0 +1,9 @@
+﻿namespace CoreLibrary.UnitTest
+{
+    public enum EErroEnum
+    {
+        TESTE_1,
+        TESTE_2,
+        TESTE_3
+    }
+}
