+++ CoreLibrary.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31025.218
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreLibrary", "CoreLibrary\CoreLibrary.csproj", "{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreLibrary.UnitTest", "CoreLibrary.UnitTest\CoreLibrary.UnitTest.csproj", "{21D325F8-9B29-42AE-BC37-124B976A30A5}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{21D325F8-9B29-42AE-BC37-124B976A30A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{21D325F8-9B29-42AE-BC37-124B976A30A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{21D325F8-9B29-42AE-BC37-124B976A30A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{21D325F8-9B29-42AE-BC37-124B976A30A5}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {380AABB2-C952-4293-BD29-BFF0FB066A4C}
+	EndGlobalSection
+EndGlobal
