--- DeadFishStudio.CoreLibrary.UnitTest/DeadFishStudio.CoreLibrary.UnitTest.csproj
@@ -1,60 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-
-    <Company>DeadFish Studio</Company>
-
-    <Authors>DeadFish Studio</Authors>
-
-    <NeutralLanguage>pt-BR</NeutralLanguage>
-
-    <PackageProjectUrl>https://github.com/luca16s/DeadFishStudio.CoreLibrary</PackageProjectUrl>
-
-    <RepositoryUrl>https://github.com/luca16s/DeadFishStudio.CoreLibrary</RepositoryUrl>
-
-    <PackageLicenseFile>LICENSE.md</PackageLicenseFile>
-
-    <Description>Core Library for DeadFish Studio Projects.</Description>
-
-    <Copyright>DeadFish Studio</Copyright>
-
-    <RunAnalyzersDuringBuild>false</RunAnalyzersDuringBuild>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AutoFixture" Version="4.15.0" />
-    <PackageReference Include="AutoFixture.Xunit2" Version="4.15.0" />
-    <PackageReference Include="Bogus" Version="33.0.2" />
-    <PackageReference Include="coverlet.collector" Version="3.0.3">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="coverlet.msbuild" Version="3.0.3">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="FluentAssertions" Version="5.10.3" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.1" />
-    <PackageReference Include="ReportGenerator" Version="4.8.7" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\DeadFishStudio.CoreLibrary\DeadFishStudio.CoreLibrary.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\LICENSE.md">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
-
-</Project>
