--- DeadFishStudio.CoreLibrary.UnitTest/EErroEnum.cs
@@ -1,9 +1,0 @@
-﻿namespace DeadFishStudio.CoreLibrary.UnitTest
-{
-    public enum EErroEnum
-    {
-        TESTE_1,
-        TESTE_2,
-        TESTE_3
-    }
-}
