--- DeadFishStudio.CoreLibrary.UnitTest/EOKEnum.cs
@@ -1,14 +1,0 @@
-﻿namespace DeadFishStudio.CoreLibrary.UnitTest
-{
-    using System.ComponentModel;
-
-    public enum EOKEnum
-    {
-        [Description("TESTE 1")]
-        TESTE_1,
-        [Description("TESTE 2")]
-        TESTE_2,
-        [Description("TESTE 3")]
-        TESTE_3
-    }
-}
