--- DeadFishStudio.CoreLibrary.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2041
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DeadFishStudio.CoreLibrary", "DeadFishStudio.CoreLibrary\DeadFishStudio.CoreLibrary.csproj", "{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeadFishStudio.CoreLibrary.UnitTest", "DeadFishStudio.CoreLibrary.UnitTest\DeadFishStudio.CoreLibrary.UnitTest.csproj", "{21D325F8-9B29-42AE-BC37-124B976A30A5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{21D325F8-9B29-42AE-BC37-124B976A30A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{21D325F8-9B29-42AE-BC37-124B976A30A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{21D325F8-9B29-42AE-BC37-124B976A30A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{21D325F8-9B29-42AE-BC37-124B976A30A5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {380AABB2-C952-4293-BD29-BFF0FB066A4C}
-	EndGlobalSection
-EndGlobal
